New: resizing console as stable as it will be
Added a lot of code. A quick summary of the bug is that locking the text in the console to the bottom while it was being resized was causing a bunch of errors. The display is stable now, and when the box is resized, the top of the text stays where it was, even when the user lets go of the mouse.

New lines in the console will jump it back to the bottom regardless of where it is.

-Also changed most CSS measurements from px to em (based off font size)
-Should add toEm and toPx to utils.js file (refs #9)
-Changed bug where consoleText div would sometimes not resize shortly after page load
-Removed some unnecessary code

Resolves #14. Finally.